fix: targeting ui dir structure
